refactor: resolve grouped filter architecture findings
