+++ .beads/interactions.jsonl
@@ -171,3 +171,5 @@
 {"id":"int-d81028f9","kind":"field_change","created_at":"2026-07-13T15:39:45.497227Z","actor":"David Ten","issue_id":"stockscreenclaude-uwl","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"Completed: hardened grouped expression invariants and UI validation, centralized filter capabilities and SQL leaf compilation, and extracted scan query routes."}}
 {"id":"int-9d83f533","kind":"field_change","created_at":"2026-07-13T16:16:33.359189Z","actor":"David Ten","issue_id":"stockscreenclaude-518","extra":{"field":"status","new_value":"in_progress","old_value":"open"}}
 {"id":"int-0032fe8d","kind":"field_change","created_at":"2026-07-13T16:52:42.023708Z","actor":"David Ten","issue_id":"stockscreenclaude-518","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"Resolved all thermo review blockers: unified canonical POST state, scan-owned grouped policy, one external decoder, and one cross-runtime field contract; relevant tests, lint, builds, and container validation complete."}}
+{"id":"int-6f72ee15","kind":"field_change","created_at":"2026-07-14T00:30:44.242222Z","actor":"David Ten","issue_id":"stockscreenclaude-giv","extra":{"field":"status","new_value":"in_progress","old_value":"open"}}
+{"id":"int-a8d7ae6f","kind":"field_change","created_at":"2026-07-14T01:00:34.801928Z","actor":"David Ten","issue_id":"stockscreenclaude-giv","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"Removed boundary dump/reparse decoding, centralized filter state, packaged the field contract, and split monolithic expression modules; relevant backend/frontend gates pass."}}
