+++ .beads/issues.jsonl
@@ -1,3 +1,4 @@
+{"id":"stockscreenclaude-giv","title":"Resolve grouped filter architecture review findings","status":"closed","priority":1,"issue_type":"task","owner":"davidten7@gmail.com","created_at":"2026-07-14T00:30:34Z","created_by":"David Ten","updated_at":"2026-07-14T01:00:35Z","started_at":"2026-07-14T00:30:44Z","closed_at":"2026-07-14T01:00:35Z","close_reason":"Removed boundary dump/reparse decoding, centralized filter state, packaged the field contract, and split monolithic expression modules; relevant backend/frontend gates pass.","dependency_count":0,"dependent_count":0,"comment_count":0}
 {"id":"stockscreenclaude-518","title":"Resolve grouped-filter thermo review blockers","description":"Collapse live scan results onto the canonical POST expression path; move grouped expression policy out of domain/common; unify API/static expression decoding; and replace duplicated browser field metadata with one shared cross-runtime contract.","acceptance_criteria":"Live quick filters and grouped filters use one POST query/export/chart path; common query has no grouped scan policy; one backend expression decoder handles API/static policies; browser aliases and builder options derive from the shared contract; relevant backend/frontend tests, lint, and build pass.","status":"closed","priority":1,"issue_type":"task","owner":"davidten7@gmail.com","created_at":"2026-07-13T16:16:27Z","created_by":"David Ten","updated_at":"2026-07-13T16:52:42Z","started_at":"2026-07-13T16:16:33Z","closed_at":"2026-07-13T16:52:42Z","close_reason":"Resolved all thermo review blockers: unified canonical POST state, scan-owned grouped policy, one external decoder, and one cross-runtime field contract; relevant tests, lint, builds, and container validation complete.","dependencies":[{"issue_id":"stockscreenclaude-518","depends_on_id":"stockscreenclaude-al4","type":"discovered-from","created_at":"2026-07-14T00:16:26Z","created_by":"David Ten","metadata":"{}"}],"dependency_count":0,"dependent_count":0,"comment_count":0}
 {"id":"stockscreenclaude-uwl","title":"Resolve grouped scan filter thermo review findings","status":"closed","priority":1,"issue_type":"task","owner":"davidten7@gmail.com","created_at":"2026-07-13T15:12:15Z","created_by":"David Ten","updated_at":"2026-07-13T15:39:45Z","started_at":"2026-07-13T15:12:21Z","closed_at":"2026-07-13T15:39:45Z","close_reason":"Completed: hardened grouped expression invariants and UI validation, centralized filter capabilities and SQL leaf compilation, and extracted scan query routes.","dependency_count":0,"dependent_count":0,"comment_count":0}
 {"id":"stockscreenclaude-i33","title":"Resolve thermo review findings for grouped scan filters","description":"Address the blocking maintainability review: dark rollout defaults, listing discovery liquidity semantics, canonical V2 presets, static expression export/parity, server-owned filter/sort capabilities, and ScanPage result lifecycle extraction.","status":"closed","priority":1,"issue_type":"task","owner":"davidten7@gmail.com","created_at":"2026-07-13T14:29:14Z","created_by":"David Ten","updated_at":"2026-07-13T14:54:24Z","started_at":"2026-07-13T14:29:19Z","closed_at":"2026-07-13T14:54:24Z","close_reason":"Resolved all thermo review findings; focused backend/frontend contracts, static export, lint, and build gates pass.","dependency_count":0,"dependent_count":0,"comment_count":0}
