+++ backend/app/domain/scanning/filter_capabilities.py
@@ -11,9 +11,7 @@ from typing import Final, Literal
 
 FilterKind = Literal["range", "categorical", "boolean", "text"]
 FieldValueType = Literal["number", "date", "string", "boolean"]
-_CONTRACT_PATH = (
-    Path(__file__).resolve().parents[4] / "contracts" / "scan_filter_fields.json"
-)
+_CONTRACT_PATH = Path(__file__).with_name("scan_filter_fields.json")
 
 
 @dataclass(frozen=True, slots=True)
@@ -26,10 +24,6 @@ class ScanFieldCapability:
     sortable: bool = False
     api_filter: bool = True
     legacy_key: str | None = None
-    builder_label: str | None = None
-    builder_category: str | None = None
-    option_source: str | None = None
-    options: tuple[str, ...] = ()
 
 
 def _default_value_type(kind: FilterKind | None) -> FieldValueType:
@@ -53,22 +47,14 @@ def _load_contract() -> tuple[ScanFieldCapability, ...]:
             raise RuntimeError("Scan filter field entries must be objects")
         field = raw.get("field")
         kind = raw.get("kind")
-        builder = raw.get("builder") or {}
         if not isinstance(field, str) or not field:
             raise RuntimeError("Scan filter fields must use non-empty names")
         if kind not in {None, "range", "categorical", "boolean", "text"}:
             raise RuntimeError(f"Unsupported scan filter kind for {field!r}")
-        if not isinstance(builder, dict):
-            raise RuntimeError(f"Builder metadata for {field!r} must be an object")
 
         value_type = raw.get("value_type", _default_value_type(kind))
         if value_type not in {"number", "date", "string", "boolean"}:
             raise RuntimeError(f"Unsupported value type for {field!r}")
-        options = builder.get("options", [])
-        if not isinstance(options, list) or any(
-            not isinstance(value, str) for value in options
-        ):
-            raise RuntimeError(f"Builder options for {field!r} must be strings")
 
         capabilities.append(
             ScanFieldCapability(
@@ -78,10 +64,6 @@ def _load_contract() -> tuple[ScanFieldCapability, ...]:
                 sortable=raw.get("sortable") is True,
                 api_filter=raw.get("api_filter") is not False,
                 legacy_key=raw.get("legacy_key"),
-                builder_label=builder.get("label"),
-                builder_category=builder.get("category"),
-                option_source=builder.get("option_source"),
-                options=tuple(options),
             )
         )
     return tuple(capabilities)
@@ -137,26 +119,6 @@ LEGACY_BOOLEAN_FILTER_FIELDS: Final = MappingProxyType(
     }
 )
 
-
-def filter_field_catalog_payload() -> list[dict[str, object]]:
-    """Serialize guided-builder metadata from the shared field contract."""
-
-    return [
-        {
-            "field": item.field,
-            "label": item.builder_label,
-            "type": item.filter_kind,
-            "value_type": item.value_type,
-            "category": item.builder_category,
-            "sortable": item.sortable,
-            "option_source": item.option_source,
-            "options": list(item.options),
-        }
-        for item in SCAN_FIELD_CAPABILITIES
-        if item.builder_label is not None
-    ]
-
-
 __all__ = [
     "ALL_FILTER_FIELD_KINDS",
     "BOOLEAN_FIELDS",
@@ -170,5 +132,4 @@ __all__ = [
     "SORT_FIELDS",
     "TEXT_FIELDS",
     "ScanFieldCapability",
-    "filter_field_catalog_payload",
 ]
