+++ backend/tests/unit/domain/test_filter_capabilities.py
@@ -4,11 +4,9 @@ from app.domain.scanning.filter_capabilities import (
     BOOLEAN_FIELDS,
     CATEGORICAL_FIELDS,
     FIELD_CAPABILITIES,
-    FILTER_FIELD_KINDS,
     RANGE_FIELDS,
     SORT_FIELDS,
     TEXT_FIELDS,
-    filter_field_catalog_payload,
 )
 from app.infra.query import feature_store_query, scan_result_query
 
@@ -28,16 +26,5 @@ def test_every_sort_field_is_supported_by_both_persistence_adapters():
     assert SORT_FIELDS <= feature_store_query.supported_sort_fields()
 
 
-def test_builder_catalog_is_a_typed_subset_of_the_server_contract():
-    catalog = filter_field_catalog_payload()
-    fields = [item["field"] for item in catalog]
-
-    assert len(fields) == len(set(fields))
-    assert set(fields) <= set(FILTER_FIELD_KINDS)
-    assert all(item["type"] == FILTER_FIELD_KINDS[item["field"]] for item in catalog)
-    assert all(item["sortable"] == (item["field"] in SORT_FIELDS) for item in catalog)
-
-    by_field = {item["field"]: item for item in catalog}
+def test_date_metadata_is_available_to_domain_validation():
     assert FIELD_CAPABILITIES["ipo_date"].value_type == "date"
-    assert by_field["rating"]["option_source"] == "ratings"
-    assert "US" in by_field["market"]["options"]
