+++ frontend/src/features/scan/filterOptions.test.js
@@ -9,7 +9,6 @@ describe('normalizeScanFilterOptions', () => {
         ibd_industries: ['Semiconductors'],
         gics_sectors: ['Technology'],
         ratings: ['Strong Buy'],
-        filter_catalog: [{ field: 'rs_rating', type: 'range' }],
       })
     ).toEqual({
       ibdIndustries: ['Semiconductors'],
@@ -20,7 +19,6 @@ describe('normalizeScanFilterOptions', () => {
         gics_sectors: ['Technology'],
         ratings: ['Strong Buy'],
       },
-      filterCatalog: [{ field: 'rs_rating', type: 'range' }],
     });
   });
 
@@ -38,7 +36,6 @@ describe('normalizeScanFilterOptions', () => {
         gics_sectors: [],
         ratings: [],
       },
-      filterCatalog: [],
     });
   });
 });
